Allow a custom nginx.conf source to be defined
In cases where configuration complexity or legacy support dictates the
need for hand-crafted configuration, allow the user to pass in a file
source for nginx.conf.

Currently it's possible to use this module and define/manage your own
virtual hosts and other configuration. This change will allow the main
config to be manually managed as well, but still take advantage of the
repo and service management of the module.